[cleanup] Extract useInvokePaletteTool
It onlys invokes the indicated tool, without being tied to a
particular UI flow.
