+++ packages/diagrams/frontend/sirius-components-diagrams/src/renderer/palette/Palette.tsx
@@ -19,15 +19,17 @@ import IconButton from '@mui/material/IconButton';
 import Paper from '@mui/material/Paper';
 import Tooltip from '@mui/material/Tooltip';
 import { Edge, Node, useStoreApi, useViewport, XYPosition } from '@xyflow/react';
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import Draggable, { DraggableData } from 'react-draggable';
 import { makeStyles } from 'tss-react/mui';
 import { EdgeData, NodeData } from '../DiagramRenderer.types';
+import { useInvokePaletteTool } from '../tools/useInvokePaletteTool';
 import {
   GQLPalette,
   GQLPaletteDivider,
   GQLPaletteEntry,
   GQLSingleClickOnDiagramElementTool,
+  GQLTool,
   GQLToolSection,
   PaletteProps,
   PaletteState,
@@ -37,7 +39,10 @@ import { PaletteQuickAccessToolBar } from './quick-access-tool/PaletteQuickAcces
 import { PaletteSearchField } from './search/PaletteSearchField';
 import { PaletteSearchResult } from './search/PaletteSearchResult';
 import { PaletteToolList } from './tool-list/PaletteToolList';
-import { usePalette } from './usePalette';
+import { useDiagramElementPalette } from './useDiagramElementPalette';
+import { useDiagramPalette } from './useDiagramPalette';
+import { usePaletteContents } from './usePaletteContents';
+import { UsePaletteContentValue } from './usePaletteContents.types';
 
 const usePaletteStyle = makeStyles<PaletteStyleProps>()((theme, props) => ({
   palette: {
@@ -126,7 +131,25 @@ export const Palette = ({
     x = (paletteX - viewportX) / viewportZoom;
     y = (paletteY - viewportY) / viewportZoom;
   }
-  const { handleToolClick, palette } = usePalette({ x, y, diagramElementId, onDirectEditClick, targetObjectId });
+
+  const { palette }: UsePaletteContentValue = usePaletteContents(diagramElementId);
+  const { hideDiagramPalette, setLastToolInvoked } = useDiagramPalette();
+  const { hideDiagramElementPalette } = useDiagramElementPalette();
+  const { invokeTool } = useInvokePaletteTool({ x, y, diagramElementId, onDirectEditClick, targetObjectId });
+
+  const closeAllPalettes = useCallback(() => {
+    hideDiagramPalette();
+    hideDiagramElementPalette();
+    domNode?.focus();
+  }, [hideDiagramPalette, hideDiagramElementPalette]);
+
+  const handleToolClick = (tool: GQLTool) => {
+    closeAllPalettes();
+    invokeTool(tool);
+    if (palette) {
+      setLastToolInvoked(palette.id, tool);
+    }
+  };
 
   useEffect(() => {
     const paletteLocation: XYPosition = computePaletteLocation(paletteX, paletteY, viewportWidth, viewportHeight);
